Added base class for NN workflows
Changed `_localWorkflow` to use it.
